initial implementation for exercise page
since we do not know anything about the query that will be
executed and we do not what data it will return, the best bet
is to hash the result as it is and compare it with the stored
solution. the solution is considered correct if it returned the same
rows and the same columns in the same order. alias names does not affect
the result. as for the hash, the requirements are to be FAST and
deterministic and xxHash satisfies these requirements.