link exercise table with user_schema table
note: breaking changes are harmless since this is still in dev.
each user schema can have many exercises and at the same time
an exercise belongs to a single category. There is no relation
that links category with schema, that is because a category is
not specific to a schema in the context of our app. This is an
sql exercise website which means categories are already
predefined e.g.: Basics, Filterling, Aggregation ... etc.
Linking the category with schema and allowing schemas
to have their own categories will introduce a huge amount
of unnecessary complexity.