add support for multiple schemas in the app
The original intention for the app was to have support for multiple
schemas where each can have their own categories and exercises.
However, due to the fact that the app was needed asap for just a single
schema, the multiple schemas support was not implemented at all because
we were short on time.