limit user access to only the current schema
The user is not allowed to access anything when executing solutions
other than the exercise's schema tables, hence this commit.
It is worth noting that if the user prefixed tables names in the
FROM clause with a schema name postgres will search for the table
in the prefixed schema, which mean relying on search path is not
secure at all but this will suffice until a better solution is
implemented.