add short_name column to user_schema
the short_name will be used to identitify a schema e.i. it will the
same as "username" for users it will be unique, identitify the schema
and used in the url, it will only allow letters. while the schema
name will be a long name that can be anything.